Initial import of wikiHow files from 15 May 2014 release.
BlueSky.skin.php is an unmodified copy of WikiHowSkin.php, and none of the
other files have been touched (yet), I've only reorganized the assets a
bit.

So far this includes files from the skins/owl directory only. There are
some files used by the skin in skins/WikiHow, but I don't want to import
that as-is because it contains a *lot* of legacy stuff that's not needed
for the BlueSky project.

Change-Id: I7d8f6329034b6349a08c7d859a737c6186d04d70