Integrated the FeaturedContributor widget by Vu Ngyuen.
FeaturedContributor was always a skin-specific feature incompatible with non-wikiHow skins, as only wikiHow skins call it and have the necessary CSS classes etc. so it makes sense to have it as a built-in feature rather than a separate extension.

Added SocialProfile avatar support to FeaturedContributor (function showFeaturedContributorWidget()), since it's more likely to be available than wikiHow's Avatar extension.

Also fixed a fatal caused by incorrect function visibility and manually set $showFeaturedArticlesSidebar to false to hide a currently empty white box.

Change-Id: Iec222ebe055a7cf5d88374a1f96ba88c9877bf5b