Integrated the essential bits and pieces of wikiHow's Notifications extension.
This is needed to prevent fatals if neither Echo nor Notifications are available.
Theoretically the orange bar of death could be removed in favor of this code, but I'm keeping it in for the time being.

Also restored the Special:Notifications link in the personal tools menu if Echo is installed.

Change-Id: Icce3cc2d7c23e05dff8ae4b1a831f9bbe1801a7a